Create TF-IDF.py
The TF-IDF algorithm:
Calculate the frequency (the number of occurrences divided by document length) for each word in your collection of documents. This is the “term frequency”, or tf

tf. (Note: Ignore punctuation and capitalization when doing this.)
Calculate how many documents each word appears in, and divide this by the total number of documents. This is the “document frequency”, or 1 / df

df. Since we wish to assign less weight to common words, we will use the inverse of this 1/df

There are different ways to combine these two numbers to assign weights to each word. The most common is the product of the term frequency and the logarithm of the inverse of the document frequency: 

tf−idf=tf×log(1÷df).